Refactored merge_pub_employees module with creation of two related modules
