+++ bmfuncts/__init__.py
@@ -17,6 +17,8 @@ from bmfuncts.use_homonyms import *
 from bmfuncts.use_otps import *
 from bmfuncts.update_employees import *
 from bmfuncts.build_pub_authors import *
+from bmfuncts.build_year_pub_empl import *
+from bmfuncts.create_hash_id import *
 from bmfuncts.merge_pub_employees import *
 from bmfuncts.consolidate_pub_list import *
 from bmfuncts.update_impact_factors import *
